delete Cuisine table, adjust Ingredient Table
