+++ db/migrate/20221027141441_create_ingredients.rb
@@ -2,7 +2,7 @@ class CreateIngredients < ActiveRecord::Migration
   def change
     create_table :ingredients do |t|
       t.string :name, :null => false, index: { unique: true, name: 'unique_ingredient' }
-      t.float :calorie_per_gram
+      t.float :calorie_per_serving
 
       t.timestamps null: false
     end
