--- db/migrate/20221027152112_add_cuisines_ref_to_recipes.rb
@@ -1,5 +1,0 @@
-class AddCuisinesRefToRecipes < ActiveRecord::Migration
-  def change
-    add_reference :recipes, :cuisine, index: true, foreign_key: true
-  end
-end
