+++ db/recipeApi.rb
@@ -17,6 +17,6 @@ response = http.request(request)
 
 resp = JSON.parse(response.read_body)
 
-File.open("rawdata/recipes.json","w") do |f|
+File.open("rawdata/recipes_1110.json","w") do |f|
     f.write(resp.to_json)
 end
